+++ service/src/test/java/org/ehrbase/service/validation/FhirTerminologyValidationTest.java
@@ -131,6 +131,29 @@ class FhirTerminologyValidationTest {
         verify(validation, never()).internalGet(Mockito.anyString());
     }
 
+    @Test
+    void testUriTagValueStripsCodeButKeepsTerminologyUrl() {
+        String basePath = "http://terminology.local";
+        Assertions.assertEquals(
+                "/fhir/CodeSystem/$validate-code?url=http://snomed.info/sct",
+                FhirTerminologyValidation.uriTagValue(
+                        basePath + "/fhir/CodeSystem/$validate-code?url=http://snomed.info/sct&code=12345"));
+
+        Assertions.assertEquals(
+                FhirTerminologyValidation.uriTagValue(
+                        basePath + "/fhir/CodeSystem/$validate-code?url=http://snomed.info/sct&code=A"),
+                FhirTerminologyValidation.uriTagValue(
+                        basePath + "/fhir/CodeSystem/$validate-code?url=http://snomed.info/sct&code=B"));
+
+        Assertions.assertNotEquals(
+                FhirTerminologyValidation.uriTagValue(
+                        basePath + "/fhir/ValueSet/$expand?url=https://hl7.org/fhir/ValueSet/icd-10"),
+                FhirTerminologyValidation.uriTagValue(
+                        basePath + "/fhir/ValueSet/$expand?url=https://hl7.org/fhir/ValueSet/administrative-gender"));
+
+        Assertions.assertEquals("/fhir/metadata", FhirTerminologyValidation.uriTagValue(basePath + "/fhir/metadata"));
+    }
+
     static ValueSet anyValueSet() {
         List<ValueSetExpansionContainsComponent> values = IntStream.range(0, 16)
                 .mapToObj(i -> anyValueSetExpansionContainsComponent())