+++ CHANGELOG.md
@@ -15,6 +15,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
  ### Changed
 - Improved template cache configuration [#1618](https://github.com/ehrbase/ehrbase/pull/1618)
 - Improved performance of template service [#1618](https://github.com/ehrbase/ehrbase/pull/1618)
+- Added metrics for external FHIR terminology server HTTP calls [#1617](https://github.com/ehrbase/ehrbase/pull/1617)
  ### Fixed 
 
 ## [2.30.1]